reorganize config settings into a new config.py file and update imports
